SaveData + Quest star ranking + Energy System
Save data between scenes and to the disk using PlayerPrefs. The current data that is saved: money, warriors hp (for the energy system), quest stars.

Quest star ranking at the end of the game and when selecting quests. Reward is based on how many more stars you get from the previous time you tried a misson (let's say you didn't play a mission and got 3 stars, then you would get the full reward; but if you already played the mission and got 1 star, and when you played it a second time you got 3 stars, you would get 2/3 of the reward on the second playthrough and 1/3 on the first one).

Added an energy system that recovers the warrior's hp when outside the battle scene.