Item functionality in battle and Item shop
Added the functionality to modify player status in battle based on equipped items.
Also added a fully functional item shop (the price will need to be edited, right now it is set to 0 in order to test, and there are no sprites for the items)

The only missing piece is saving the bought and equiped equipment to the disk and changing shop so that you can't buy them again if you already bought them.